# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAPRINA Sales Operations System API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  received: "received",
  delayed: "delayed",
  returned: "returned",
  partial_received: "partial_received",
} as const;

export interface Order {
  id: number;
  customerName: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  address: string | null;
  product: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  status: OrderStatus;
  /** @nullable */
  partialQuantity: number | null;
  /** @nullable */
  shippingCompanyId: number | null;
  /** @nullable */
  productId: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderBody {
  customerName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  product: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  unitPrice: number;
  /** @nullable */
  shippingCompanyId?: number | null;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type UpdateOrderBodyStatus =
  (typeof UpdateOrderBodyStatus)[keyof typeof UpdateOrderBodyStatus];

export const UpdateOrderBodyStatus = {
  pending: "pending",
  received: "received",
  delayed: "delayed",
  returned: "returned",
  partial_received: "partial_received",
} as const;

export interface UpdateOrderBody {
  customerName?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  product?: string;
  /** @minimum 1 */
  quantity?: number;
  /** @minimum 0 */
  unitPrice?: number;
  status?: UpdateOrderBodyStatus;
  /** @nullable */
  partialQuantity?: number | null;
  /** @nullable */
  shippingCompanyId?: number | null;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface OrdersSummary {
  totalOrders: number;
  pendingOrders: number;
  receivedOrders: number;
  delayedOrders: number;
  returnedOrders: number;
  partialOrders: number;
  totalRevenue: number;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  sku: string | null;
  totalQuantity: number;
  reservedQuantity: number;
  soldQuantity: number;
  lowStockThreshold: number;
  unitPrice: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductBody {
  name: string;
  /** @nullable */
  sku?: string | null;
  /** @minimum 0 */
  totalQuantity: number;
  /** @minimum 0 */
  lowStockThreshold?: number;
  /** @minimum 0 */
  unitPrice: number;
}

export interface UpdateProductBody {
  name?: string;
  /** @nullable */
  sku?: string | null;
  /** @minimum 0 */
  totalQuantity?: number;
  /** @minimum 0 */
  lowStockThreshold?: number;
  /** @minimum 0 */
  unitPrice?: number;
}

export interface ShippingCompany {
  id: number;
  name: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  website: string | null;
  /** @nullable */
  notes: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateShippingCompanyBody {
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  notes?: string | null;
  isActive?: boolean;
}

export interface UpdateShippingCompanyBody {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  notes?: string | null;
  isActive?: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  search?: string;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  received: "received",
  delayed: "delayed",
  returned: "returned",
  partial_received: "partial_received",
} as const;