# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: CAPRINA Sales Operations System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Order management operations
  - name: products
    description: Inventory management
  - name: shipping
    description: Shipping companies
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, received, delayed, returned, partial_received]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/summary:
    get:
      operationId: getOrdersSummary
      tags: [orders]
      summary: Get orders summary statistics
      responses:
        "200":
          description: Orders summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersSummary"
  /orders/recent:
    get:
      operationId: getRecentOrders
      tags: [orders]
      summary: Get 5 most recent orders
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products in inventory
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shipping-companies:
    get:
      operationId: listShippingCompanies
      tags: [shipping]
      summary: List all shipping companies
      responses:
        "200":
          description: List of shipping companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShippingCompany"
    post:
      operationId: createShippingCompany
      tags: [shipping]
      summary: Create a shipping company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShippingCompanyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShippingCompany"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shipping-companies/{id}:
    patch:
      operationId: updateShippingCompany
      tags: [shipping]
      summary: Update a shipping company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShippingCompanyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShippingCompany"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteShippingCompany
      tags: [shipping]
      summary: Delete a shipping company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        product:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        totalPrice:
          type: number
          format: float
        status:
          type: string
          enum: [pending, received, delayed, returned, partial_received]
        partialQuantity:
          type: ["integer", "null"]
        shippingCompanyId:
          type: ["integer", "null"]
        productId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - phone
        - address
        - product
        - quantity
        - unitPrice
        - totalPrice
        - status
        - partialQuantity
        - shippingCompanyId
        - productId
        - notes
        - createdAt
        - updatedAt
    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        product:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: float
          minimum: 0
        shippingCompanyId:
          type: ["integer", "null"]
        productId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - customerName
        - product
        - quantity
        - unitPrice
    UpdateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        product:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: float
          minimum: 0
        status:
          type: string
          enum: [pending, received, delayed, returned, partial_received]
        partialQuantity:
          type: ["integer", "null"]
        shippingCompanyId:
          type: ["integer", "null"]
        productId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    OrdersSummary:
      type: object
      properties:
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        receivedOrders:
          type: integer
        delayedOrders:
          type: integer
        returnedOrders:
          type: integer
        partialOrders:
          type: integer
        totalRevenue:
          type: number
          format: float
      required:
        - totalOrders
        - pendingOrders
        - receivedOrders
        - delayedOrders
        - returnedOrders
        - partialOrders
        - totalRevenue
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: ["string", "null"]
        totalQuantity:
          type: integer
        reservedQuantity:
          type: integer
        soldQuantity:
          type: integer
        lowStockThreshold:
          type: integer
        unitPrice:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sku
        - totalQuantity
        - reservedQuantity
        - soldQuantity
        - lowStockThreshold
        - unitPrice
        - createdAt
        - updatedAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        sku:
          type: ["string", "null"]
        totalQuantity:
          type: integer
          minimum: 0
        lowStockThreshold:
          type: integer
          minimum: 0
        unitPrice:
          type: number
          format: float
          minimum: 0
      required:
        - name
        - totalQuantity
        - unitPrice
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        sku:
          type: ["string", "null"]
        totalQuantity:
          type: integer
          minimum: 0
        lowStockThreshold:
          type: integer
          minimum: 0
        unitPrice:
          type: number
          format: float
          minimum: 0
    ShippingCompany:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - website
        - notes
        - isActive
        - createdAt
    CreateShippingCompanyBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - name
    UpdateShippingCompanyBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]