# lib/api-zod/src/generated/types/createShippingCompanyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAPRINA Sales Operations System API
 * OpenAPI spec version: 0.2.0
 */

export interface CreateShippingCompanyBody {
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  notes?: string | null;
  isActive?: boolean;
}