# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAPRINA Sales Operations System API
 * OpenAPI spec version: 0.2.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  customerName: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  address: string | null;
  product: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  status: OrderStatus;
  /** @nullable */
  partialQuantity: number | null;
  /** @nullable */
  shippingCompanyId: number | null;
  /** @nullable */
  productId: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
  updatedAt: Date;
}