# lib/api-zod/src/generated/types/updateOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAPRINA Sales Operations System API
 * OpenAPI spec version: 0.2.0
 */
import type { UpdateOrderBodyStatus } from "./updateOrderBodyStatus";

export interface UpdateOrderBody {
  customerName?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  product?: string;
  /** @minimum 1 */
  quantity?: number;
  /** @minimum 0 */
  unitPrice?: number;
  status?: UpdateOrderBodyStatus;
  /** @nullable */
  partialQuantity?: number | null;
  /** @nullable */
  shippingCompanyId?: number | null;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  notes?: string | null;
}