# lib/api-zod/src/generated/types/updateProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAPRINA Sales Operations System API
 * OpenAPI spec version: 0.2.0
 */

export interface UpdateProductBody {
  name?: string;
  /** @nullable */
  sku?: string | null;
  /** @minimum 0 */
  totalQuantity?: number;
  /** @minimum 0 */
  lowStockThreshold?: number;
  /** @minimum 0 */
  unitPrice?: number;
}